cursada/bimestre3/Backend/clases/clase 41 -  1 de julio de 2022/clinica-odontologica/src/main/java/com/dh/clinica/service/PacienteService.java
package com.dh.clinica.service;


import com.dh.clinica.exceptions.ResourceNotFoundException;
import com.dh.clinica.repository.IDao;
import com.dh.clinica.model.Paciente;
import org.springframework.stereotype.Service;

import java.util.Date;
import java.util.List;
import java.util.Optional;
@Service
public class PacienteService {

    private IDao<Paciente> pacienteIDao;

    public PacienteService(IDao<Paciente> pacienteIDao) {
        this.pacienteIDao = pacienteIDao;
    }

    public Paciente guardar(Paciente p) {
        p.setFechaIngreso(new Date());
        return pacienteIDao.guardar(p);
    }

    public Optional<Paciente> buscar(Integer id) {
        return pacienteIDao.buscar(id);
    }

    public List<Paciente> buscarTodos() {
        return pacienteIDao.buscarTodos();
    }

    public void eliminar(Integer id) throws ResourceNotFoundException{

        if(buscar(id).isEmpty())
            throw  new ResourceNotFoundException("No existe paciente con id: "+id);
        pacienteIDao.eliminar(id);
    }

    public Paciente actualizar(Paciente p) {
        return pacienteIDao.actualizar(p);
    }
}
